Test error Can't call method 'address' ofundefined
Test error Can't call method 'address' ofundefined. When executing grunt
test we get this error:

"TypeError: Cannot call method 'address' of undefined". This is because
we are defining at the begining of the testApi.spec.js this:

var app = require('../server/app').app,
    api = supertest(app);

This means that supertest is trying to connect but we still haven't type
any code in the server/app.js file

Contact: joaquin@isla.io